+++ src/ExtendedPromiseInterface.php
@@ -6,5 +6,6 @@ interface ExtendedPromiseInterface extends PromiseInterface
 {
     public function done(callable $onFulfilled = null, callable $onRejected = null, callable $onProgress = null);
     public function otherwise(callable $onRejected);
+    public function always(callable $onFulfilledOrRejected);
     public function progress(callable $onProgress);
 }