+++ src/Promise.php
@@ -64,6 +64,19 @@ class Promise implements ExtendedPromiseInterface
         });
     }
 
+    public function always(callable $onFulfilledOrRejected)
+    {
+        return $this->then(function($value) use ($onFulfilledOrRejected) {
+            return resolve($onFulfilledOrRejected())->then(function () use ($value) {
+                return $value;
+            });
+        }, function ($reason) use ($onFulfilledOrRejected) {
+            return resolve($onFulfilledOrRejected())->then(function () use ($reason) {
+                return new RejectedPromise($reason);
+            });
+        });
+    }
+
     public function progress(callable $onProgress)
     {
         return $this->then(null, null, $onProgress);